fix: tsc compiler and refactoring batch process
