+++ firebase.json
@@ -1,0 +1,14 @@
+{
+  "emulators": {
+    "firestore": {
+      "port": 8080
+    },
+    "database": {
+      "port": 9000
+    },
+    "ui": {
+      "enabled": true,
+      "port": 9090
+    }
+  }
+}
