+++ package.json
@@ -33,17 +33,21 @@
     "@commitlint/cli": "^9.1.1",
     "@commitlint/config-conventional": "^9.1.1",
     "@types/algoliasearch": "^3.34.10",
-    "@typescript-eslint/eslint-plugin": "^3.6.1",
-    "@typescript-eslint/parser": "^3.6.1",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
+    "babel-jest": "^26.3.0",
     "commander": "^5.1.0",
     "cz-conventional-changelog": "^3.2.0",
-    "eslint": "^7.4.0",
+    "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "husky": "^4.2.5",
-    "prettier": "^2.0.5",
     "firebase-admin": "^9.0.0",
-    "standard-version": "^8.0.2"
+    "firebase-tools": "^8.7.0",
+    "husky": "^4.2.5",
+    "jest": "^26.4.0",
+    "prettier": "^2.0.5",
+    "standard-version": "^8.0.2",
+    "ts-jest": "^26.2.0"
   },
   "dependencies": {
     "algoliasearch": "^4.3.0",
@@ -52,8 +56,8 @@
   },
   "peerDependencies": {
     "algoliasearch": "^4.3.0",
-    "typescript": "^3.9.6",
-    "firebase-admin": "^9.0.0"
+    "firebase-admin": "^9.0.0",
+    "typescript": "^3.9.6"
   },
   "husky": {
     "hooks": {
